Updated dependencies and CI to comply with new MSRV policy.
The "Minimum Supported Rust Version" (MSRV) states that
without any features enabled the MSRV is 1.34. This is the base MSRV.
If the 'module_disambiguation' feature is enabled, the MSRV increases to 1.42.
The rest of the features have the base MSRV.
Therefore, since 'module_diambiguation' is on by default, the default
MSRV is 1.42.

Most of this crate's dependencies don't have an MSRV, meaning
we will have to depend on specific versions to ensure our MSRV
is not broken by a new version of a dependency.