next_token: eagerly load two tokens instead of one
While this technically changes the semantics of the code, it shouldn't
impact the overall behaviour of the method as a *normal* implementation
of Iterator would be idempotent once the end reached. As such, if the
first token is None, we can safely discard the second one.

This change is done in order to avoid using a reference to `in_group`
within the guard clause, as Rust 1.34 complains about it. It is probable
that the more recent versions accept it from 1.39 on, see https://github.com/rust-lang/rust/blob/master/RELEASES.md#language-7