Relaxed MSRV to not include dependencies.
This allows us to be less strict with pinning dependencies
since we just need a minimal version to work.
This will allow users to depend on us even if they
have a different version requirement to one of the crates
that we use too.

CI now tests that minimal versions work on the MSRV.