Duplicate's edition no longer leaks into user code.
`proc_macro::Span` tracks the edition of the code.
Previously, duplicate's edition was inserted into the expanded
code by not setting the span of created `proc_macro::Group`s.

Added one specific test (in tests/no_features/edition_2021)
that exhibits the problem, however, it is difficult to know
if there are other cases.
The test `ensure_no_group_new` was added to try and catch
any wrong creation of groups in the future.