Switched to using 'duplicate_macrotest' crate instead of 'macrotest'
'duplicate_macrotest' is a fork of 'macrotest' that contains
changes that allows testing for failure (a feature not released
in 'macrotest' at this time). This will allow us to test 'duplicate's
error messages.

It will also allow us to avoid new versions of 'macrotest' that
might break the builds even though the major version wasn't incremented.
This has happened multiple times, so we cannot trust the versioning of
'macrotest' to be compatible with semantic versioning.